Cease and desist destructive tendencies (#24)
I'm hitting a problem when restarting nexus, where the `registry` suddenly changes blobstore and repos vanish.

The problem is that `nexus-setup.sh` runs as an `ExecPostStart` script on the systemd service, and when it runs it has two destructive options.

These destructive options make it impossible to restart the nexus service without harming an existing registry. For example, in Fawkes we create a registry in the fawkes blobstore. Restarting nexus causes that registry to be recreated again under the default blobstore, this ends up breaking image pulls.

Signed-off-by: Russell Bunch <doomslayer@hpe.com>